style custom-tooltip
