+++ gt7dashboard/tab_manager.py
@@ -6,7 +6,6 @@ from .tabs.config_tab import ConfigTab
 from .tabs.fuel_tab import FuelTab
 from .tabs.laptime_analysis_tab import LapTimeAnalysisTab
 from .tabs.racetime_datatable_tab import RaceTimeDataTableTab
-from . import gt7diagrams
 
 class TabManager:
     """Manages all tabs in the GT7 Dashboard"""
@@ -14,36 +13,9 @@ class TabManager:
     def __init__(self, app_instance):
         self.app = app_instance
 
-        # Create shared components
-        self.race_diagram = gt7diagrams.RaceDiagram(width=1000)
-
-        # Create race line figure
-        race_line_tooltips = [("index", "$index"), ("Brakepoint", "")]
-        race_line_width = 250
-
-        self.s_race_line = figure(
-            title="Race Line",
-            x_axis_label="x",
-            y_axis_label="z",
-            match_aspect=True,
-            width=race_line_width,
-            height=race_line_width,
-            active_drag="box_zoom",
-            tooltips=race_line_tooltips,
-        )
-        # We set this to true, since maps appear flipped in the game
-        # compared to their actual coordinates
-        self.s_race_line.y_range.flipped = True
-        self.s_race_line.toolbar.autohide = True
-
         # Create tabs
         self.race_lines_tab = RaceLinesTab(app_instance)
-        
         self.race_tab = RaceTab(app_instance)
-        self.race_tab.set_diagrams(self.race_diagram, self.s_race_line)
-        self.race_tab.initialize()
-        self.race_tab.finalize_layout()
-
         self.config_tab = ConfigTab(app_instance)
         self.fuel_tab = FuelTab(app_instance)
         self.laptime_table_tab = LapTimeAnalysisTab(app_instance)
